- Sources splitted to several files, cmake now finds them with GLOB - readFile() function added
