+++ src/shaders/shader.frag
@@ -4,6 +4,7 @@
 layout(location = 0) in vec3 fragColor;
 layout(location = 0) out vec4 outColor;
 
-void main() {
+void main()
+{
     outColor = vec4(fragColor, 1.0);
 }
