+++ src/utils/io.h
@@ -1,0 +1,11 @@
+#ifndef HELLO_VULKAN_IO_H
+#define HELLO_VULKAN_IO_H
+
+#include <vector>
+#include <xstring>
+
+using namespace std;
+
+static vector<char> readFile(const string& filename);
+
+#endif //HELLO_VULKAN_IO_H
